# contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Quicker.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Create a workspace and its owner account",
        "operationId": "postAuthSignup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Password sign-in; may answer with a tenant choice or an MFA challenge",
        "operationId": "postAuthLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/select-tenant": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthSelectTenant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Complete an MFA challenge with a TOTP code, a recovery code or a WebAuthn assertion",
        "operationId": "postAuthMfaVerify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/webauthn/options": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthMfaWebauthnOptions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaWebAuthnOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnAssertionOptionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll/totp": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Start TOTP enrolment when the workspace policy requires MFA before the first session",
        "operationId": "postAuthMfaEnrollTotp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaWebAuthnOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll/totp/confirm": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthMfaEnrollTotpConfirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpConfirmWithChallenge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate a refresh token",
        "operationId": "postAuthRefresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password/forgot": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthPasswordForgot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/auth/invitations/accept": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthInvitationsAccept",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sso/{tenantSlug}/{connectionCode}/start": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Redirects to the tenant's identity provider",
        "operationId": "getAuthSsoByTenantSlugByConnectionCodeStart",
        "parameters": [
          {
            "name": "tenantSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connectionCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnTo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/auth/sso/exchange": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "postAuthSsoExchange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "My account"
        ],
        "summary": "Who am I, with effective permissions",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "My account"
        ],
        "operationId": "patchMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/password": {
      "post": {
        "tags": [
          "My account"
        ],
        "operationId": "postMePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/me/logout": {
      "post": {
        "tags": [
          "My account"
        ],
        "operationId": "postMeLogout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/me/step-up": {
      "post": {
        "tags": [
          "My account"
        ],
        "summary": "Re-prove identity for sensitive actions",
        "operationId": "postMeStepUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "My account"
        ],
        "operationId": "deleteMeSessionsBySessionId",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/me/mfa/totp/confirm": {
      "post": {
        "tags": [
          "My account"
        ],
        "operationId": "postMeMfaTotpConfirm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SelectTenantRequest": {
        "required": [
          "challengeToken",
          "tenantSlug"
        ],
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string"
          },
          "tenantSlug": {
            "type": "string"
          }
        }
      },
...
      "SignupRequest": {
        "required": [
          "tenantName",
          "slug",
          "ownerEmail",
          "ownerName",
          "password"
        ],
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "ownerEmail": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "default": "en"
          }
        }
      },
...
      "TokenResponse": {
        "required": [
          "accessToken",
          "refreshToken",
          "expiresInSeconds",
          "tenant",
          "user",
          "membershipId"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresInSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantSummary"
          },
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}